API

Modules

LegendSpecFits.LegendSpecFits

Types and constants

Functions and macros

LegendSpecFits.estimate_single_peak_stats

LegendSpecFits.ex_gauss_pdf

LegendSpecFits.expand_vars

LegendSpecFits.gamma_peakshape

LegendSpecFits.gauss_pdf

LegendSpecFits.hist_loglike

LegendSpecFits.step_gauss

LegendSpecFits.weibull_from_mx

Documentation

LegendSpecFits.LegendSpecFits — Module

LegendSpecFits

Template for Julia packages.

LegendSpecFits.estimate_single_peak_stats — Method

estimate_single_peak_stats(h::Histogram)

Estimate statistics/parameters for a single peak in the given histogram h.

h must only contain a single peak. The peak should have a Gaussian-like shape.

Returns a NamedTuple with the fields

* `peak_pos`: estimated position of the peak (in the middle of the peak)
* `peak_fwhm`: full width at half maximum (FWHM) of the peak
* `peak_sigma`: estimated standard deviation of the peak
* `peak_counts`: estimated number of counts in the peak
* `mean_background`: estimated mean background value

LegendSpecFits.ex_gauss_pdf — Method

ex_gauss_pdf(x::Real, μ::Real, σ::Real, θ::Real)

The PDF of an Exponentially modified Gaussian distribution with Gaussian parameters μ, σ and exponential scale θ at x.

It is the PDF of the distribution that descibes the random process rand(Normal(μ, σ)) + rand(Exponential(θ)).

LegendSpecFits.expand_vars — Method

expand_vars(v::NamedTuple)::StructArray

Expand all fields in v (scalars or arrays) to same array size and return a StructArray.

LegendSpecFits.gamma_peakshape — Method

gamma_peakshape(
    x::Real, μ::Real, σ::Real, n::Real,
    step_amplitude::Real, skew_fraction::Real, skew_width::Real
)

Describes the shape of a typical gamma peak in a detector.

LegendSpecFits.gauss_pdf — Method

LegendSpecFits.gauss_pdf(x::Real, μ::Real, σ::Real)

Equivalent to pdf(Normal(μ, σ), x)

LegendSpecFits.hist_loglike — Method

hist_loglike(f_fit::Base.Callable, h::Histogram{<:Real,1})

Calculate the Poisson log-likelihood of a fit function f_fit(x) and a histogram h. f_fit must accept all values x on the horizontal axis of the histogram.

Currently uses a simple midpoint-rule integration of f_fit over the bins of h.

LegendSpecFits.step_gauss — Method

step_gauss(x::Real, μ::Real, σ::Real)

Evaluates the convulution of a Heaviside step function and the PDF of Normal(μ, σ) at x.

The result does not correspond to a PDF as it is not normalizable.

LegendSpecFits.weibull_from_mx — Function

weibull_from_mx(m::Real, x::Real, p_x::Real = 0.6827)::Weibull

Construct a Weibull distribution with a given median m and a given p_x-quantile x.

Useful to construct priors for positive quantities.