API

Modules

LegendSpecFits.LegendSpecFits

Types and constants

Functions and macros

LegendSpecFits.aoe_compton_background_peakshape

LegendSpecFits.aoe_compton_peakshape

LegendSpecFits.aoe_compton_signal_peakshape

LegendSpecFits.array_to_tuple

LegendSpecFits.autocal_energy

LegendSpecFits.background_peakshape

LegendSpecFits.calibrate_energy!

LegendSpecFits.correct_aoe!

LegendSpecFits.ctc_energy

LegendSpecFits.cut_single_peak

LegendSpecFits.estimate_combined_peak_stats

LegendSpecFits.estimate_fwhm

LegendSpecFits.estimate_single_peak_stats

LegendSpecFits.estimate_single_peak_stats_psd

LegendSpecFits.ex_gauss_pdf

LegendSpecFits.ex_step_gauss

LegendSpecFits.expand_vars

LegendSpecFits.exponential_decay

LegendSpecFits.f_optimize_ctc

LegendSpecFits.fit_aoe_compton

LegendSpecFits.fit_aoe_corrections

LegendSpecFits.fit_calibration

LegendSpecFits.fit_enc_sigmas

LegendSpecFits.fit_fwhm

LegendSpecFits.fit_fwhm_ft_fep

LegendSpecFits.fit_half_centered_trunc_gauss

LegendSpecFits.fit_peaks

LegendSpecFits.fit_peaks_combined

LegendSpecFits.fit_single_aoe_compton

LegendSpecFits.fit_single_peak_th228

LegendSpecFits.fit_single_trunc_gauss

LegendSpecFits.gamma_peakshape

LegendSpecFits.gauss_pdf

LegendSpecFits.generate_aoe_compton_bands

LegendSpecFits.get_continuum_surrival_fraction

LegendSpecFits.get_dep_peakhists

LegendSpecFits.get_distribution_transform

LegendSpecFits.get_mc_value_shapes

LegendSpecFits.get_n_after_psd_cut

LegendSpecFits.get_peak_fwhm_th228

LegendSpecFits.get_peak_surrival_fraction

LegendSpecFits.get_peaks_surrival_fractions

LegendSpecFits.get_psd_cut

LegendSpecFits.hist_loglike

LegendSpecFits.linear_function

LegendSpecFits.lowEtail_peakshape

LegendSpecFits.signal_peakshape

LegendSpecFits.simple_calibration

LegendSpecFits.step_gauss

LegendSpecFits.subhist

LegendSpecFits.tuple_to_array

LegendSpecFits.weibull_from_mx

Documentation

LegendSpecFits.LegendSpecFits — Module

LegendSpecFits

Template for Julia packages.

LegendSpecFits.aoe_compton_background_peakshape — Method

aoe_compton_background_peakshape(
    x::Real, μ::Real, σ::Real,
    background::Real, δ::Real
)

Describes the background shape of a typical A/E Compton peak in a detector as a step like background for MSE events.

LegendSpecFits.aoe_compton_peakshape — Method

aoe_compton_peakshape(
    x::Real, μ::Real, σ::Real, n::Real,
    background::Real, δ::Real
)

Describes the shape of a typical A/E Compton peak in a detector as a gaussian SSE peak and a step like background for MSE events.

LegendSpecFits.aoe_compton_signal_peakshape — Method

aoe_compton_signal_peakshape(
    x::Real, μ::Real, σ::Real, n::Real
)

Describes the signal shape of a typical A/E Compton peak in a detector as a gaussian SSE peak.

LegendSpecFits.array_to_tuple — Method

array_to_tuple(a::AbstractArray, as_nt::NamedTuple)

Return a NamedTuple with the values of a in the order given by fieldnames(as_nt).

LegendSpecFits.autocal_energy — Method

autocal_energy(E_raw::AbstractArray{<:Real})

Compute an energy calibration from raw reconstructed energy deposition values.

LegendSpecFits.background_peakshape — Method

background_peakshape(
    x::Real, μ::Real, σ::Real, n::Real,
    skew_fraction::Real, skew_width::Real,
)

Describes the background part of the shape of a typical gamma peak in a detector.

LegendSpecFits.calibrate_energy! — Method

calibrate_energy!(e::AbstractArray{<:Real}, pars::PropDict)

Calibrate energy values in-place.

LegendSpecFits.correct_aoe! — Method

correctAoE!(aoe::Array{T}, e::Array{T}, aoe_corrections::NamedTuple{(:e, :μ, :f_μ_scs, :μ_scs_slope, :μ_scs_intercept, :σ, :σ_scs, :f_σ_scs)}) where T<:Real

Correct the AoE values in the aoe array using the corrections in aoe_corrections.

LegendSpecFits.ctc_energy — Method

ctc_energy(e::Array{T}, qdrift::Array{T}, peak::T, window::T) where T<:Real

Correct for the drift time dependence of the energy by minimizing the ratio of the FWHM and the peak height of the peak around peak in e with a cut window of window. The drift time dependence is given by qdrift.

Returns

* `peak`: peak position
* `window`: window size
* `fct`: correction factor
* `bin_width`: optimal bin width

LegendSpecFits.cut_single_peak — Method

cut_single_peak(x::Array, min_x::Float64, max_x::Float64, n_bins::Int=15000, relative_cut::Float64=0.5)

Cut out a single peak from the array x between min_x and max_x. The number of bins is the number of bins to use for the histogram. The relative cut is the fraction of the maximum counts to use for the cut.

Returns

* `max`: maximum position of the peak
* `low`: lower edge of the cut peak
* `high`: upper edge of the cut peak

LegendSpecFits.estimate_combined_peak_stats — Method

estimate_combined_peak_stats(peakstats::StructArray,; calib_type::Symbol=:th228)

Estimate the peak position, FWHM, sigma, counts and background of a peak from a histogram.

LegendSpecFits.estimate_fwhm — Method

estimate_fwhm(v::NamedTuple, v_err::NamedTuple)

Get the FWHM of a peak from the fit parameters.

Returns

* `fwhm`: the FWHM of the peak

LegendSpecFits.estimate_single_peak_stats — Method

estimate_single_peak_stats(h::Histogram, calib_type::Symbol=:th228)

Estimate statistics/parameters for a single peak in the given histogram h.

h must only contain a single peak. The peak should have a Gaussian-like shape. calib_type specifies the calibration type. Currently only :th228 is implemented. If you want get the peak statistics for a PSD calibration, use :psd.

Returns

NamedTuple with the fields * peak_pos: estimated position of the peak (in the middle of the peak) * peak_fwhm: full width at half maximum (FWHM) of the peak * peak_sigma: estimated standard deviation of the peak * peak_counts: estimated number of counts in the peak * mean_background: estimated mean background value

LegendSpecFits.estimate_single_peak_stats_psd — Method

estimate_single_peak_stats_psd(h::Histogram{T}) where T<:Real

Estimate peak parameters for a single peak in a histogram using the maximum, the FWHM and the area of the peak.

Returns

* `peak_pos`: Position of the peak
* `peak_fwhm`: Full width at half maximum of the peak
* `peak_sigma`: Standard deviation of the peak
* `peak_counts`: Counts of the peak
* `mean_background`: Mean background of the peak

LegendSpecFits.ex_gauss_pdf — Method

ex_gauss_pdf(x::Real, μ::Real, σ::Real, θ::Real)

The PDF of an Exponentially modified Gaussian distribution with Gaussian parameters μ, σ and exponential scale θ at x.

It is the PDF of the distribution that descibes the random process rand(Normal(μ, σ)) + rand(Exponential(θ)).

LegendSpecFits.ex_step_gauss — Method

ex_step_gauss(x::Real, l::Real, k::Real, t::Real, d::Real)

Evaluates an extended step gauss model at x with parameters l, k, t and d.

LegendSpecFits.expand_vars — Method

expand_vars(v::NamedTuple)::StructArray

Expand all fields in v (scalars or arrays) to same array size and return a StructArray.

LegendSpecFits.exponential_decay — Method

exponential_decay(x::Real, amplitude::Real, decay::Real, offset::Real)

Evaluates an exponential decay function at x with parameters amplitude, decay and offset.

LegendSpecFits.f_optimize_ctc — Method

f_optimize_ctc(fct, e, qdrift, bin_width)

Calculate the ratio of the FWHM and the peak height of the peak around peak in e with a cut window of window. The drift time dependence is given by e_ctc = e + fct * qdrift.

Returns

* `fwhm / p_height`: FWHM of the peak divided by peak height

LegendSpecFits.fit_aoe_compton — Method

fit_aoe_compton(peakhists::Array, peakstats::StructArray, compton_bands::Array{T}) where T<:Real

Fit the A/E Compton bands using the f_aoe_compton function consisting of a gaussian SSE peak and a step like background for MSE events.

Returns

* `result`: Dict of NamedTuples of the fit results containing values and errors for each compton band
* `report`: Dict of NamedTuples of the fit report which can be plotted for each compton band

LegendSpecFits.fit_aoe_corrections — Method

fit_aoe_corrections(e::Array{<:Real}, μ::Array{T}, σ::Array{T}) where T<:Real

Fit the corrections for the AoE value of the detector.

LegendSpecFits.fit_calibration — Method

fitCalibration

Fit the calibration lines to a linear function.

Returns

* `slope`: the slope of the linear fit
* `intercept`: the intercept of the linear fit

LegendSpecFits.fit_enc_sigmas — Method

fit_enc_sigmas(enc_grid::Matrix{T}, enc_grid_rt::StepRangeLen{Quantity{<:T}, Base.TwicePrecision{Quantity{<:T}}, Base.TwicePrecision{Quantity{<:T}}, Int64}, min_enc::T, max_enc::T, nbins::Int64, rel_cut_fit::T) where T<:Real

Fit the ENC values in enc_grid for each RT in enc_grid_rt with a Gaussian and return the optimal RT and the corresponding ENC value.

Arguments

enc_grid: 2D array of ENC values for each RT in enc_grid_rt

enc_grid_rt: 1D array of RT values for which the ENC values in enc_grid are calculated

min_enc: minimum ENC value to consider for the fit

max_enc: maximum ENC value to consider for the fit

nbins: number of bins to use for the histogram of ENC values

rel_cut_fit: relative cut value to use for the fit

Returns

rt: optimal RT value

min_enc: corresponding ENC value

LegendSpecFits.fit_fwhm — Method

fitFWHM(fit_fwhm(peaks::Vector{T}, fwhm::Vector{T}) where T<:Real

Fit the FWHM of the peaks to a quadratic function.

Returns

* `qbb`: the FWHM at 2039 keV
* `err`: the uncertainties of the fit parameters
* `v`: the fit result parameters
* `f_fit`: the fitted function

LegendSpecFits.fit_fwhm_ft_fep — Method

fit_fwhm_ft_fep(e_grid::Matrix{T}, e_grid_ft::StepRangeLen{Quantity{<:T}, Base.TwicePrecision{Quantity{<:T}}, Base.TwicePrecision{Quantity{<:T}}, Int64}) where T <:Real

Fit the FWHM values in e_grid for each FT in e_grid_ft with a Gamma Peakshape and return the optimal FT and the corresponding FWHM value.

Arguments

e_grid: 2D array of energy values for each FT in e_grid_ft

e_grid_ft: 1D array of FT values for which the FWHM values in e_grid are calculated

Returns

ft: optimal FT value

min_fwhm: corresponding FWHM value

LegendSpecFits.fit_half_centered_trunc_gauss — Method

fit_half_centered_trunc_gauss(x::Array, cuts::NamedTuple{(:low, :high, :max), Tuple{Float64, Float64, Float64}})

Fit a single truncated Gaussian to the data x between cut.low and cut.high. The peak center is fixed at μ and the peak is cut in half either in the left or right half.

Returns report and result` with:

* `f_fit`: fitted function
* `μ`: mean of the Gaussian
* `μ_err`: error of the mean
* `σ`: standard deviation of the Gaussian
* `σ_err`: error of the standard deviation
* `n`: number of counts in the peak

LegendSpecFits.fit_peaks — Method

fitPeaks

Perform a fit of the peakshape to the data in peakhists using the initial values in peakstats to the calibration lines in th228_lines.

Returns

* `peak_fit_plots`: array of plots of the peak fits
* `return_vals`: dictionary of the fit results

LegendSpecFits.fit_peaks_combined — Method

fit_peaks_combined(peakhists::Array, peakstats::StructArray, th228_lines::Array{T},; calib_type::Symbol=:th228, uncertainty::Bool=true, fixed_position::Bool=false) where T<:Real

Fit the peaks of a histogram to a combined peakshape function while sharring parameters between peaks.

LegendSpecFits.fit_single_aoe_compton — Method

fit_single_aoe_compton(h::Histogram, ps::NamedTuple{(:peak_pos, :peak_fwhm, :peak_sigma, :peak_counts, :mean_background, :μ, :σ), NTuple{7, T}}; uncertainty::Bool=true) where T<:Real

Perform a fit of the peakshape to the data in h using the initial values in ps while using the f_aoe_compton function consisting of a gaussian SSE peak and a step like background for MSE events.

Returns

* `result`: NamedTuple of the fit results containing values and errors
* `report`: NamedTuple of the fit report which can be plotted

LegendSpecFits.fit_single_peak_th228 — Method

fit_single_peak_th228(h::Histogram, ps::NamedTuple{(:peak_pos, :peak_fwhm, :peak_sigma, :peak_counts, :mean_background), NTuple{5, T}};, uncertainty::Bool=true, fixed_position::Bool=false) where T<:Real

Perform a fit of the peakshape to the data in h using the initial values in ps while using the gamma_peakshape with low-E tail. Also, FWHM is calculated from the fitted peakshape with MC error propagation. The peak position can be fixed to the value in ps by setting fixed_position=true.

Returns

* `result`: NamedTuple of the fit results containing values and errors
* `report`: NamedTuple of the fit report which can be plotted

LegendSpecFits.fit_single_trunc_gauss — Method

fit_single_trunc_gauss(x::Array, cuts::NamedTuple{(:low, :high, :max), Tuple{Float64, Float64, Float64}})

Fit a single truncated Gaussian to the data x between min_x and max_x. Returns report and resultwith: *ffit: fitted function *μ: mean of the Gaussian *μerr: error of the mean *σ: standard deviation of the Gaussian *σ_err: error of the standard deviation *n`: number of counts in the peak

LegendSpecFits.gamma_peakshape — Method

gamma_peakshape(
    x::Real, μ::Real, σ::Real, n::Real,
    step_amplitude::Real, skew_fraction::Real, skew_width::Real,
    background::Real
)

Describes the shape of a typical gamma peak in a detector with a flat background.

LegendSpecFits.gauss_pdf — Method

LegendSpecFits.gauss_pdf(x::Real, μ::Real, σ::Real)

Equivalent to pdf(Normal(μ, σ), x)

LegendSpecFits.generate_aoe_compton_bands — Method

generate_aoe_compton_bands(aoe::Array{<:Real}, e::Array{<:Real}, compton_bands::Array{<:Real}, compton_window::T) where T<:Real

Generate histograms for the A/E Compton bands and estimate peak parameters. The compton bands are cutted out of the A/E spectrum and then binned using the Freedman-Diaconis Rule. For better performance the binning is only done in the area around the peak. The peak parameters are estimated using the estimate_single_peak_stats_psd function.

Returns

* `peakhists`: Array of histograms for each compton band
* `peakstats`: StructArray of peak parameters for each compton band
* `min_aoe`: Array of minimum A/E values for each compton band
* `max_aoe`: Array of maximum A/E values for each compton band

LegendSpecFits.get_continuum_surrival_fraction — Method

get_continuum_surrival_fraction(aoe::Array{T}, e::Array{T}, center::T, window::T, psd_cut::T,; uncertainty=true) where T<:Real

Get the surrival fraction of a continuum after a PSD cut value psd_cut for a given center and window size.

Returns

center: Center of the continuum

window: Window size

n_before: Number of counts before the cut

n_after: Number of counts after the cut

sf: Surrival fraction

err: Uncertainties

LegendSpecFits.get_dep_peakhists — Method

get_aoe_peakhists(aoe::Array{T}, e::Array{T}) where T<:Real

Get the histograms of the DEP in the AoE spectrum.

Returns

hist: Histogram of the DEP peak

stats: Stats of the DEP peak

dep: Position of the DEP peak

window: Window size around the DEP peak

e: Energy values in the window around the DEP peak

aoe: AoE values in the window around the DEP peak

LegendSpecFits.get_distribution_transform — Function

get_distribution_transform(d::Distribution, pprior::Prior)

Return a DistributionTransform for the given Distribution and Prior.

LegendSpecFits.get_mc_value_shapes — Method

get_mc_value_shapes(v::NamedTuple, v_err::NamedTuple, n::Int64)

Return a NamedTuple with the same fields as v and v_err but with Normal distributions for each field.

LegendSpecFits.get_n_after_psd_cut — Method

get_n_after_psd_cut(psd_cut::T, aoe::Array{T}, e::Array{T}, peak::T, window::T,; uncertainty=true) where T<:Real

Get the number of counts after a cut value psd_cut for a given peak and window size whiile performing a peak fit with fixed position.

# Returns

n: Number of counts after the cut

n_err: Uncertainty of the number of counts after the cut

LegendSpecFits.get_peak_fwhm_th228 — Function

get_peak_fwhm_th228(v_ml::NamedTuple, v_ml_err::NamedTuple)

Get the FWHM of a peak from the fit parameters while performing a MC error propagation.

Returns

* `fwhm`: the FWHM of the peak
* `fwhm_err`: the uncertainty of the FWHM of the peak

LegendSpecFits.get_peak_surrival_fraction — Method

get_peak_surrival_fraction(aoe::Array{T}, e::Array{T}, peak::T, window::T, psd_cut::T,; uncertainty=true) where T<:Real

Get the surrival fraction of a peak after a PSD cut value psd_cut for a given peak and window size whiile performing a peak fit with fixed position.

Returns

peak: Peak position

n_before: Number of counts before the cut

n_after: Number of counts after the cut

sf: Surrival fraction

err: Uncertainties

LegendSpecFits.get_peaks_surrival_fractions — Method

get_peaks_surrival_fractions(aoe::Array{T}, e::Array{T}, peaks::Array{T}, peak_names::Array{Symbol}, windows::Array{T}, psd_cut::T,; uncertainty=true) where T<:Real

Get the surrival fraction of a peak after a PSD cut value psd_cut for a given peak and window size while performing a peak fit with fixed position.

Return

result: Dict of results for each peak

report: Dict of reports for each peak

LegendSpecFits.get_psd_cut — Method

get_psd_cut(aoe::Array{T}, e::Array{T},; dep_sf::Float64=0.9) where T<:Real

Get the PSD cut value for a given DEP surrival fraction dep_sf (Default: 90%). The cut value is determined by finding the cut value where the number of counts after the cut is dep_sf of the number of counts before the cut. The algorithm is based on a root search function and expecting a Bisection.

Returns

cut: PSD cut value

n0: Number of counts before the cut

n90: Number of counts after the cut

LegendSpecFits.hist_loglike — Method

hist_loglike(f_fit::Base.Callable, h::Histogram{<:Real,1})

Calculate the Poisson log-likelihood of a fit function f_fit(x) and a histogram h. f_fit must accept all values x on the horizontal axis of the histogram.

Currently uses a simple midpoint-rule integration of f_fit over the bins of h.

LegendSpecFits.linear_function — Method

linear_function(x::Real, slope::Real, intercept::Real)

Evaluates a linear function at x with parameters slope and intercept.

LegendSpecFits.lowEtail_peakshape — Method

lowEtail_peakshape(
    x::Real, μ::Real, σ::Real, n::Real,
    skew_fraction::Real, skew_width::Real,
)

Describes the low-E signal tail part of the shape of a typical gamma peak in a detector.

LegendSpecFits.signal_peakshape — Method

signal_peakshape(
    x::Real, μ::Real, σ::Real, n::Real,
    skew_fraction::Real, skew_width::Real,
)

Describes the signal part of the shape of a typical gamma peak in a detector.

LegendSpecFits.simple_calibration — Method

simple_calibration(e_uncal::Array, th228_lines::Array, window_size::Float64=25.0, n_bins::Int=15000, calib_type::String="th228")

Perform a simple calibration for the uncalibrated energy array e_uncal using the calibration type calib_type and the calibration lines th228_lines. The window size is the size of the window around the calibration line to use for the calibration. The number of bins is the number of bins to use for the histogram.

Returns * h_calsimple: histogram of the calibrated energy array * h_uncal: histogram of the uncalibrated energy array * c: calibration factor * fep_guess: estimated full energy peak (FEP) * peakhists: array of histograms around the calibration lines * peakstats: array of statistics for the calibration line fits

LegendSpecFits.step_gauss — Method

step_gauss(x::Real, μ::Real, σ::Real)

Evaluates the convulution of a Heaviside step function and the PDF of Normal(μ, σ) at x.

The result does not correspond to a PDF as it is not normalizable.

LegendSpecFits.subhist — Method

subhist(h::Histogram, r::Tuple{<:Real,<:Real})

Return a new Histogram with the bins in the range r.

LegendSpecFits.tuple_to_array — Method

tuple_to_array(nt::NamedTuple, fields::Vector{Symbol})

Return an array with the values of the fields in nt in the order given by fields.

LegendSpecFits.weibull_from_mx — Function

weibull_from_mx(m::Real, x::Real, p_x::Real = 0.6827)::Weibull

Construct a Weibull distribution with a given median m and a given p_x-quantile x.

Useful to construct priors for positive quantities.